refactor: pro system with ProSpec pattern
Introduce a flexible ProSpec pattern to support multiple problem
types (Batch, Communication, TwoStep, OutputOnly) in preparation
for CMS integration.

Architecture Changes:
- Add ProType enum and BaseConfig/BatchConfig classes
- Implement ProSpec abstract pattern with BatchProblemSpec
- Store problem-specific config in JSON (problem.config field)
- Store testdata files in JSON (testdata.files field)

Handler Modularization:
- Split monolithic handler into 7 focused modules:
  general, add, update, judge, limit, subtask, filemanager
- Implement dispatcher pattern for type-specific operations:
  submit, code, judge, testdata
- All handlers route through unified paths with dispatching

Database Migrations:
- Add problem_type and config columns to problem table
- Migrate existing data to BatchConfig JSON format
- Replace testdata inputfile/outputfile with files JSON

Template Organization:
- Create prospec/batch/ template structure
- Use dynamic {% include %} based on problem_type
- Prepare structure for future problem types

This refactoring maintains backward compatibility while enabling
easy addition of new problem types. Each type can have its own
config schema, handlers, and templates.
