+++ src/handlers/code.py
@@ -1,36 +1,58 @@
+"""
+Unified code display handler.
+Dispatches to problem type-specific handlers.
+"""
 import tornado.escape
 
 from handlers.base import RequestHandler, reqenv
 from services.code import CodeService
-from services.chal import Compiler
+from services.chal import Compiler, ChalService
+from services.pro import ProService, ProType, ProConst
 
 
 class CodeHandler(RequestHandler):
+    """Unified code display handler - dispatches to type-specific handlers."""
+
     @reqenv
     async def get(self):
         self.error(('Eacces', 'Permission denied'))
 
     @reqenv
     async def post(self):
+        """Handle code display request - dispatch to type-specific handler."""
         chal_id = int(self.get_argument('chal_id'))
 
-        err, code, compiler_type = await CodeService.inst.get_code(chal_id, self.acct)
+        # Get challenge to determine problem type
+        err, chal = await ChalService.inst.get_chal(chal_id, self.acct)
         if err:
             return self.error(err)
 
-        if compiler_type in [Compiler.GCC, Compiler.CLANG, Compiler.GPP, Compiler.CLANGPP]:
-            compiler_type = 'cpp'
-        elif compiler_type == Compiler.RUST:
-            compiler_type = 'rust'
-        elif compiler_type == Compiler.PYTHON3:
-            compiler_type = 'python'
-        elif compiler_type == Compiler.JAVA:
-            compiler_type = 'java'
-        else:
-            compiler_type = 'cpp'
+        # Determine allowed pro statuses based on contest and user type
+        allow_statuses = ProConst.PRO_STATUS_NORMAL_USER
+        if chal.contest_id:
+            allow_statuses = ProConst.PRO_STATUS_CONTEST_USER
+        elif self.acct.is_kernel():
+            allow_statuses = ProConst.PRO_STATUS_KERNEL_USER
 
-        res = {
-            'compiler_type': compiler_type,
-            'code': tornado.escape.xhtml_escape(code),
-        }
-        self.error(('S', res))
+        err, pro = await ProService.inst.get_pro(chal.pro_id, allow_statuses)
+        if err:
+            return self.error(err)
+
+        # Dispatch to problem type-specific handler
+        if pro.problem_type == ProType.BATCH:
+            from handlers.prospec.batch.code import BatchCodeHandler
+            handler = BatchCodeHandler(self.application, self.request, db=self.db, rs=self.rs)
+            handler.acct = self.acct
+            handler._transforms = []
+            return await handler.post()
+        elif pro.problem_type == ProType.COMMUNICATION:
+            # Future: Communication code display
+            return self.error(('Enotsupport', 'Communication problem type not yet supported'))
+        elif pro.problem_type == ProType.TWOSTEP:
+            # Future: TwoStep code display (might show two code sections)
+            return self.error(('Enotsupport', 'TwoStep problem type not yet supported'))
+        elif pro.problem_type == ProType.OUTPUTONLY:
+            # Future: OutputOnly code display (show uploaded files)
+            return self.error(('Enotsupport', 'OutputOnly problem type not yet supported'))
+        else:
+            return self.error(('Eparam', 'Invalid problem type'))
