+++ src/handlers/manage/pro/subtask.py
@@ -1,0 +1,156 @@
+from dataclasses import asdict
+from collections import defaultdict
+
+from handlers.base import RequestHandler, reqenv, require_permission
+from services.log import LogService
+from services.pro import ProService, ProConst, SubtaskConfig
+from services.user import UserConst
+from utils.numeric import parse_str_to_list
+
+ALLOW_STATUSES = [ProConst.STATUS_ONLINE, ProConst.STATUS_CONTEST, ProConst.STATUS_HIDDEN]
+
+
+class ManageProSubtaskHandler(RequestHandler):
+    @reqenv
+    @require_permission(UserConst.ACCTTYPE_KERNEL)
+    async def get(self):
+        pro_id = int(self.get_argument('proid'))
+        err, pro = await ProService.inst.get_pro(pro_id, ALLOW_STATUSES)
+        if err:
+            return self.error(err)
+
+        await self.render('manage/pro/updatesubtask', page='pro', pro_id=pro_id, config=pro.config)
+
+    @reqenv
+    @require_permission(UserConst.ACCTTYPE_KERNEL)
+    async def post(self):
+        reqtype = self.get_argument('reqtype')
+        pro_id = int(self.get_argument('pro_id'))
+
+        err, pro = await ProService.inst.get_pro(pro_id, ALLOW_STATUSES)
+        if err:
+            return self.error(err)
+
+        if reqtype == "updaterate":
+            subtask_id = int(self.get_argument('subtask'))
+            rate = int(self.get_argument('rate'))
+
+            subtask_configs = pro.config.subtask_configs
+            if subtask_id not in subtask_configs:
+                return self.error(('Enoext', 'Subtask not found'))
+
+            subtask_configs[subtask_id].rate = rate
+            await ProService.inst.update_pro_config(pro_id, pro.problem_type, pro.config)
+            await LogService.inst.add_log(
+                f'{self.acct.name} has sent a request to update rate of subtask#{subtask_id} for problem #{pro_id}',
+                'manage.pro.update.subtask.updaterate',
+                {'rate': rate}
+            )
+            self.error(('S', ''))
+
+        elif reqtype == "setdepsubtasks":
+            subtask_id = int(self.get_argument('subtask'))
+            dep_subtasks = set(map(lambda x: x - 1, parse_str_to_list(self.get_argument('dep_subtasks'))))
+
+            subtask_configs = pro.config.subtask_configs
+            if subtask_id not in subtask_configs:
+                return self.error(('Enoext', 'Subtask not found'))
+
+            subtask_configs[subtask_id].dependency_subtasks = set(dep_subtasks)
+
+            if self.have_cycle(subtask_configs):
+                return self.error(('Eparam', 'Dependency subtasks have cycle'))
+
+            await ProService.inst.update_pro_config(pro_id, pro.problem_type, pro.config)
+            await LogService.inst.add_log(
+                f'{self.acct.name} has sent a request to set dependency subtasks to subtask#{subtask_id} for problem #{pro_id}',
+                'manage.pro.update.subtask.setdepsubtasks',
+                {'dependency_subtasks': list(dep_subtasks)}
+            )
+            self.error(('S', ''))
+
+        elif reqtype == "addsubtask":
+            rate = int(self.get_argument('rate'))
+
+            subtask_configs = pro.config.subtask_configs
+            subtask_configs[len(subtask_configs)] = SubtaskConfig(len(subtask_configs), [], set(), rate)
+
+            await ProService.inst.update_pro_config(pro_id, pro.problem_type, pro.config)
+            await LogService.inst.add_log(
+                f'{self.acct.name} has sent a request to add a new subtask for problem #{pro_id}',
+                'manage.pro.update.subtask.addsubtask',
+                {'rate': rate, 'subtask_id': len(subtask_configs) - 1}
+            )
+            self.error(('S', ''))
+
+        elif reqtype == 'deletesubtask':
+            subtask_id = int(self.get_argument('subtask'))
+
+            subtask_configs = pro.config.subtask_configs
+            if subtask_id not in subtask_configs:
+                return self.error(('Enoext', 'Subtask not found'))
+
+            subtask_configs.pop(subtask_id)
+            remain_subtasks = list(subtask_configs.values())
+            subtask_configs.clear()
+
+            for new_subtask_id, subtask in enumerate(remain_subtasks):
+                subtask_configs[new_subtask_id] = subtask
+
+            await ProService.inst.update_pro_config(pro_id, pro.problem_type, pro.config)
+            await LogService.inst.add_log(
+                f'{self.acct.name} has sent a request to delete a subtask for problem #{pro_id}',
+                'manage.pro.update.subtask.deletesubtask',
+            )
+            self.error(('S', ''))
+
+        elif reqtype == 'settestdata':
+            subtask_id = int(self.get_argument('subtask'))
+            testdatas = parse_str_to_list(self.get_argument('testdatas'))
+
+            subtask_configs = pro.config.subtask_configs
+            if subtask_id not in subtask_configs:
+                return self.error(('Enoext', 'Subtask not found'))
+
+            subtask_configs[subtask_id].testdatas.clear()
+            for testdata_id in testdatas:
+                if testdata_id not in pro.config.testdatas:
+                    continue
+                subtask_configs[subtask_id].testdatas.append(pro.config.testdatas[testdata_id])
+
+            await ProService.inst.update_pro_config(pro_id, pro.problem_type, pro.config)
+            await LogService.inst.add_log(
+                f'{self.acct.name} has sent a request to set testdatas to subtask#{subtask_id} for problem #{pro_id}',
+                'manage.pro.update.subtask.settestdata',
+                {
+                    'testdatas': [testdata.testdata_id for testdata in subtask_configs[subtask_id].testdatas]
+                }
+            )
+            self.error(('S', ''))
+
+    def have_cycle(self, subtask_configs: dict[int, SubtaskConfig]) -> bool:
+        vis = defaultdict(int)
+        graph = defaultdict(list)
+
+        for u, sconfig in subtask_configs.items():
+            for v in sconfig.dependency_subtasks:
+                graph[u].append(v)
+
+        def dfs(u) -> bool:
+            nonlocal vis
+            vis[u] = 1
+            for v in graph[u]:
+                if vis[v] == 1:
+                    return False
+                elif vis[v] == 0:
+                    if not dfs(v):
+                        return False
+            vis[u] = 2
+            return True
+
+        for u in subtask_configs.keys():
+            if vis[u] == 0:
+                if not dfs(u):
+                    return True
+
+        return False
