+++ src/handlers/prospec/__init__.py
@@ -1,0 +1,1 @@
+# ProSpec handlers for different problem types
