+++ src/services/prospec/__init__.py
@@ -1,0 +1,4 @@
+from services.prospec.base import ProSpec
+from services.prospec.batch import BatchProblemSpec
+
+__all__ = ['ProSpec', 'BatchProblemSpec']
