+++ src/static/templ/manage/pro/update.html
@@ -419,7 +419,7 @@
 		<button type="button" class="btn btn-primary submit">Update</button>
     </form>
 
-	<a class="btn btn-primary" href="{{ url(f'/manage/pro/updatetests/?proid={ pro.pro_id }') }}">Edit Subtask</a>
+	<a class="btn btn-primary" href="{{ url(f'/manage/pro/updatesubtask/?proid={ pro.pro_id }') }}">Edit Subtask</a>
 	<a class="btn btn-primary" href="{{ url(f'/manage/pro/updatetestdata/?proid={ pro.pro_id }') }}">Edit Testdata</a>
     <a class="btn btn-primary" href="{{ url(f'/manage/pro/filemanager/?proid={ pro.pro_id }') }}">File Manager</a>
 </div>
