+++ src/static/templ/manage/pro/updatesubtask.html
@@ -19,7 +19,7 @@
 					return;
 				}
 
-                $.post(`{{ url('/be/manage/pro/updatetests') }}`, {
+                $.post(`{{ url('/be/manage/pro/updatesubtask') }}`, {
 					'reqtype': 'updaterate',
 					'pro_id': pro_id,
 					'subtask': subtask,
@@ -44,7 +44,7 @@
 
         $("button.add-subtask").on('click', function(e) {
             let rate = parseInt(prompt('New subtask rate: '));
-            $.post(`{{ url('/be/manage/pro/updatetests') }}`, {
+            $.post(`{{ url('/be/manage/pro/updatesubtask') }}`, {
                 'reqtype': 'addsubtask',
                 'pro_id': pro_id,
                 'rate': rate,
@@ -68,7 +68,7 @@
         document.querySelectorAll('button.delete-subtask').forEach(el => {
             el.addEventListener('click', function(e) {
                 let subtask = $(this).parents('.accordion-collapse').attr('subtask');
-				$.post(`{{ url('/be/manage/pro/updatetests') }}`, {
+				$.post(`{{ url('/be/manage/pro/updatesubtask') }}`, {
 					'reqtype': 'deletesubtask',
 					'pro_id': pro_id,
                     'subtask': subtask
@@ -99,7 +99,7 @@
                     return;
                 }
 
-				$.post(`{{ url('/be/manage/pro/updatetests') }}`, {
+				$.post(`{{ url('/be/manage/pro/updatesubtask') }}`, {
 					'reqtype': 'settestdata',
 					'pro_id': pro_id,
                     'subtask': subtask,
@@ -131,7 +131,7 @@
                     return;
                 }
 
-				$.post(`{{ url('/be/manage/pro/updatetests') }}`, {
+				$.post(`{{ url('/be/manage/pro/updatesubtask') }}`, {
 					'reqtype': 'setdepsubtasks',
 					'pro_id': pro_id,
                     'subtask': subtask,
@@ -182,8 +182,8 @@
                     <button class="btn btn-danger delete-subtask">Delete Current Subtask</button>
 
                     <div class="input-group mb-3">
-                        <input id="weight" class="form-control mb-1" type="number" value="{{ subtask_config.rate }}" placeholder="Weight">
-                        <button class="btn btn-primary edit-weight">Update Rate</button>
+                        <input id="rate" class="form-control mb-1" type="number" value="{{ subtask_config.rate }}" placeholder="Rate">
+                        <button class="btn btn-primary edit-rate">Update Rate</button>
                     </div>
 
                     <label for="testdatas" class="form-label">Testdata id. Starts from 0, for examples: <code>0</code>, <code>1-4</code>, <code>2,4,9-11,13-18</code></label>
