+++ src/static/templ/prospec/batch/code.html
@@ -1,0 +1,17 @@
+{# Batch problem code display - Simple pre/code structure for AJAX loading #}
+<pre id="code" class="line-numbers"
+    data-language=""
+    data-prismjs-copy="複製" data-prismjs-copy-error="Copy Failed!!!" data-prismjs-copy-success="Copy Success!!!">
+    <code class=""></code>
+</pre>
+
+    </div>
+</div>
+
+<script>
+    // Syntax highlighting will be applied by Prism.js when this template is included
+    if (typeof Prism !== 'undefined') {
+        Prism.highlightAll();
+    }
+</script>
+
