+++ src/static/templ/prospec/batch/submit.html
@@ -1,0 +1,103 @@
+{% from services.chal import COMPILER_INFOS %}
+<link rel="stylesheet" type="text/css" href="{{ url('/src/submit.css') }}">
+
+<script type="text/javascript">
+    function init() {
+        var j_submit = $('#submit');
+        let pro_id = "{{ pro.pro_id }}";
+        let contest_id = "{{ contest_id }}";
+
+        j_submit.find('input.file').on('change', function(e) {
+            var reader = new FileReader();
+
+            reader.onload = function(e) {
+                j_submit.find('textarea.code').val(reader.result);
+            };
+
+            reader.readAsText(this.files[0]);
+        });
+
+        j_submit.find('button.submit').on('click', function(e) {
+            let code = j_submit.find('textarea.code').val();
+            let compiler_type = j_submit.find('#compilerList').val();
+
+            if (code.trim() === '') {
+                index.show_notify_dialog('Code cannot be empty', index.DIALOG_TYPE.error);
+                return;
+            }
+
+            let url = '';
+            if (contest_id !== "0") {
+                url = `{{ url('/be/contests/${contest_id}/submit') }}`;
+            } else {
+                url = `{{ url('/be/submit') }}`;
+            }
+
+            $.post(url, {
+                'reqtype': 'submit',
+                'pro_id' : pro_id,
+                'code': code,
+                'compiler_type': compiler_type,
+            }, function(res) {
+                res = JSON.parse(res);
+                if (res.status == 'S') {
+                    if (contest_id !== "0") {
+                        url = `{{ url('/contests/${contest_id}/chal/${res.data}/') }}`;
+                    } else {
+                        url = `{{ url('/chal/${res.data}/') }}`;
+                    }
+
+                    index.go(url);
+                } else {
+                    index.show_notify_dialog(res.data, index.DIALOG_TYPE.error);
+                }
+            });
+        });
+    }
+</script>
+
+<div class="row">
+    <div class="col-lg-2"></div>
+    <div id="submit" class="col-lg-8">
+        <h3>{{ pro.pro_id }} / {{ pro.name }}</h3>
+
+        <div class="alert alert-info" role="alert">
+            <strong>Batch Problem</strong> - Submit your source code for traditional batch judging
+        </div>
+
+        <form>
+            <div class="mb-3">
+                <label for="compilerList" class="form-label">Compiler</label>
+                <select id="compilerList" class="form-select">
+                    {% for compiler in allow_compilers %}
+                        <option value="{{ compiler }}" {% if user.last_compiler == compiler %} selected {% end %}>
+                            {{ COMPILER_INFOS[compiler].version_name }}
+                        </option>
+                    {% end %}
+                </select>
+                <div class="form-text">Select the compiler for your source code</div>
+            </div>
+
+            <div class="mb-3">
+                <label for="codeArea" class="form-label">Source Code</label>
+                <textarea id="codeArea" class="code form-control" placeholder="Paste your code here..." rows="20"></textarea>
+                <div class="form-text">Or use the file upload below to load code from a file</div>
+            </div>
+
+            <div class="mb-3">
+                <label for="fileInput" class="form-label">Upload File (Optional)</label>
+                <input type="file" id="fileInput" class="file form-control">
+                <div class="form-text">Select a source code file to automatically fill the code area</div>
+            </div>
+
+            <div class="d-grid gap-2">
+                <button type="button" class="btn btn-primary btn-lg submit">
+                    <i class="bi bi-upload"></i> Submit Code
+                </button>
+            </div>
+        </form>
+
+    </div>
+    <div class="col-lg-2"></div>
+</div>
+
